+++ src/image/stringBlitter.cpp
@@ -52,8 +52,8 @@ StringImageInfo StringBlitter::evaluateTextImageSize(const std::string &str) {
     FT_GlyphSlot slot = face->glyph;
     const FT_Bitmap &bitmap = face->glyph->bitmap;
 
-    int imgWidth = 0, imgHeight = 0;
-    int maxBearingY = 0;
+    unsigned int imgWidth = 0, imgHeight = 0;
+    unsigned int maxBearingY = 0;
 
     for (unsigned int i = 0; i < str.size(); i++) {
         // retrieve glyph index
@@ -66,8 +66,8 @@ StringImageInfo StringBlitter::evaluateTextImageSize(const std::string &str) {
         CHECK_FREETYPE_ERROR(FT_Render_Glyph(slot, FT_RENDER_MODE_NORMAL));
 
         imgWidth += slot->advance.x >> 6;
-        imgHeight = std::max(imgHeight, slot->bitmap_top + bitmap.rows);
-        maxBearingY = std::max(maxBearingY, slot->bitmap_top);
+        imgHeight = std::max(imgHeight, static_cast<unsigned int>(slot->bitmap_top + bitmap.rows));
+        maxBearingY = std::max(maxBearingY, static_cast<unsigned int>(slot->bitmap_top));
     }
     
     log4cpp::log_console->debugStream() << "[FreeType] Image Size (in pixels): " << imgWidth << "x" << imgHeight;
@@ -84,13 +84,16 @@ Image StringBlitter::generateTextImageGraylevel(const std::string &str) {
     const StringImageInfo imgInfo = evaluateTextImageSize(str);
 
     Image image(imgInfo.imgWidth, imgInfo.imgHeight, 1u);
-    int x = 0;
+    unsigned int x = 0;
     
     for (unsigned int i = 0; i < str.size(); i++) {
         index = FT_Get_Char_Index(face, static_cast<unsigned char>(str[i])); 
         CHECK_FREETYPE_ERROR(FT_Load_Glyph(face, index, FT_LOAD_DEFAULT));
         CHECK_FREETYPE_ERROR(FT_Render_Glyph(slot, FT_RENDER_MODE_NORMAL));
-        blitCharacter(image, x+slot->bitmap_left, imgInfo.maxBearingY - slot->bitmap_top, bitmap);
+        blitCharacter(image, 
+                x+static_cast<unsigned int>(slot->bitmap_left), 
+                static_cast<unsigned int>(static_cast<int>(imgInfo.maxBearingY) - slot->bitmap_top),
+                bitmap);
         x += slot->advance.x >> 6;
     }
 
@@ -105,9 +108,9 @@ Image StringBlitter::generateTextImageRGBA(const std::string &str, const ColorRG
     
     const unsigned char *colorData = reinterpret_cast<const unsigned char *>(&color);
 
-    for (int j = 0; j < rgbaImg.height; j++) {
-        for (int i = 0; i < rgbaImg.width; i++) {
-            for (int k = 0; k < channels; k++) {
+    for (unsigned int j = 0; j < rgbaImg.height; j++) {
+        for (unsigned int i = 0; i < rgbaImg.width; i++) {
+            for (unsigned int k = 0; k < channels; k++) {
                 float color = static_cast<float>(colorData[k])*grayscaleImg.data[j*grayscaleImg.width+i]/255.0f;
                 rgbaImg.data[(j*rgbaImg.width+i)*channels+k] = static_cast<unsigned char>(color);
             }
@@ -118,7 +121,7 @@ Image StringBlitter::generateTextImageRGBA(const std::string &str, const ColorRG
     return rgbaImg;
 }
         
-void StringBlitter::blitCharacter(const Image &image, int x, int y, const FT_Bitmap &bitmap) {
+void StringBlitter::blitCharacter(const Image &image, unsigned int x, unsigned int y, const FT_Bitmap &bitmap) {
     // check if bitmap has 256 gray levels
     assert(bitmap.pixel_mode == FT_PIXEL_MODE_GRAY);
     assert(bitmap.num_grays  == 0x100);
@@ -128,9 +131,9 @@ void StringBlitter::blitCharacter(const Image &image, int x, int y, const FT_Bit
 
     unsigned char *glyphData = bitmap.buffer;
     
-    for (int j = 0; j < bitmap.rows; j++) {
-        for (int i = 0; i < bitmap.width; i++) {
-            image.data[(y+j)*image.width+(x+i)] = glyphData[j*bitmap.pitch +i];
+    for (unsigned int j = 0; j < static_cast<unsigned int>(bitmap.rows); j++) {
+        for (unsigned int i = 0; i < static_cast<unsigned int>(bitmap.width); i++) {
+            image.data[(y+j)*image.width+(x+i)] = glyphData[j*static_cast<unsigned int>(bitmap.pitch) +i];
         }
     }
 }
